requirements.txt
category==0.0.3
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2
django-3-jet==1.0.8
django-ckeditor==6.2.0
django-dynamic-preferences==1.12.0
django-environ==0.9.0
django-js-asset==2.0.0
django-parler==2.3
django-rosetta==0.9.8
idna==3.3
persisting-theory==0.2.1
Pillow
polib==1.1.1
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
django-extensions==3.2.0
django-recaptcha==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
whitenoise==6.5.0
